Classification/CNN-Based/requirements.txt
torch
torchvision
Pillow
matplotlib
gradio
numpy